# Only load the `isEmail` function from the validator library