--- a/lib/checks/010-package-json.js
+++ b/lib/checks/010-package-json.js
@@ -1,6 +1,6 @@
 const _ = require('lodash');
 const semver = require('semver');
-const validator = require('validator');
+const isEmail = require('validator/lib/isEmail');
 const _private = {};
 const packageJSONFileName = 'package.json';
 const versions = require('../utils').versions;
@@ -92,7 +92,7 @@ _private.validatePackageJSONFields = function validatePackageJSONFields(packageJ
         markFailed('versionIsSemverCompliant');
     }
 
-    if (packageJSON.author && packageJSON.author.email && !validator.isEmail(packageJSON.author.email)) {
+    if (packageJSON.author && packageJSON.author.email && !isEmail(packageJSON.author.email)) {
         markFailed('authorEmailIsValid');
     }