issue https://github.com/TryGhost/Team/issues/706

This reduces parsing time as node doesn't need to parse the whole validator.js library.

## Benchmark before the change

```
hyperfine --warmup 3  --min-runs 30 '/Users/thibautpatel/.nvm/versions/node/v14.17.0/bin/node bin/cli.js ../Casper'
Benchmark #1: /Users/thibautpatel/.nvm/versions/node/v14.17.0/bin/node bin/cli.js ../Casper
  Time (mean ± σ):     464.4 ms ±   8.3 ms    [User: 491.5 ms, System: 79.2 ms]
  Range (min … max):   453.6 ms … 495.5 ms    30 runs
```

## Benchmark after the change

```
Benchmark #1: /Users/thibautpatel/.nvm/versions/node/v14.17.0/bin/node bin/cli.js ../Casper
  Time (mean ± σ):     443.8 ms ±   6.0 ms    [User: 478.0 ms, System: 74.3 ms]
  Range (min … max):   433.5 ms … 461.6 ms    30 runs
```

So this is approximately 20ms shaved from an unzipped gscan check.